app/auth-service/requirements.txt
Flask==3.1.0
flask-cors==5.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
Werkzeug==3.1.0
gunicorn==23.0.0
